+++ Application/Items/Edit.cs
@@ -1,4 +1,5 @@
-﻿using Domain.Entities;
+﻿using AutoMapper;
+using Domain.Entities;
 using MediatR;
 using Persistence;
 using System;
@@ -8,27 +9,29 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Application.Activities
+namespace Application.Items
 {
-    public class Delete
+    public class Edit
     {
         public class Command : IRequest
         {
-            public Guid Id { get; set; }
+            public Item Item { get; set; }
         }
         public class Handler : IRequestHandler<Command>
         {
             private readonly ItemDbContext _context;
+            private readonly IMapper _mapper;
 
-            public Handler(ItemDbContext context)
+            public Handler(ItemDbContext context, IMapper mapper)
             {
                 _context = context;
+                _mapper = mapper;
             }
 
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
-                var item = await _context.Items.FindAsync(request.Id);
-                _context.Remove(item);
+                var item = await _context.Items.FindAsync(request.Item.Id);
+                _mapper.Map(request.Item, item);
                 await _context.SaveChangesAsync();
                 return Unit.Value;
             }
